replace install.py with new-install.py
